# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * ajTech API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface SuccessResponse {
  success: boolean;
}

export interface Project {
  id: string;
  title: string;
  description: string;
  /** @nullable */
  longDescription?: string | null;
  /** @nullable */
  imageUrl?: string | null;
  tags: string[];
  category: string;
  /** @nullable */
  liveUrl?: string | null;
  /** @nullable */
  githubUrl?: string | null;
  featured: boolean;
  order: number;
  createdAt: string;
  updatedAt: string;
}

export interface ProjectInput {
  /** @minLength 1 */
  title: string;
  /** @minLength 1 */
  description: string;
  longDescription?: string;
  imageUrl?: string;
  tags?: string[];
  category: string;
  liveUrl?: string;
  githubUrl?: string;
  featured?: boolean;
  order?: number;
}

export interface ProjectUpdate {
  title?: string;
  description?: string;
  longDescription?: string;
  imageUrl?: string;
  tags?: string[];
  category?: string;
  liveUrl?: string;
  githubUrl?: string;
  featured?: boolean;
  order?: number;
}

export interface BlogPost {
  id: string;
  title: string;
  slug: string;
  /** @nullable */
  excerpt?: string | null;
  content: string;
  /** @nullable */
  coverImageUrl?: string | null;
  tags: string[];
  published: boolean;
  /** @nullable */
  readTime?: number | null;
  createdAt: string;
  updatedAt: string;
}

export interface BlogPostInput {
  /** @minLength 1 */
  title: string;
  slug?: string;
  excerpt?: string;
  content: string;
  coverImageUrl?: string;
  tags?: string[];
  published?: boolean;
  readTime?: number;
}

export interface BlogPostUpdate {
  title?: string;
  slug?: string;
  excerpt?: string;
  content?: string;
  coverImageUrl?: string;
  tags?: string[];
  published?: boolean;
  readTime?: number;
}

export interface Service {
  id: string;
  title: string;
  description: string;
  icon: string;
  features: string[];
  order: number;
}

export interface ServiceInput {
  /** @minLength 1 */
  title: string;
  /** @minLength 1 */
  description: string;
  icon: string;
  features?: string[];
  order?: number;
}

export interface ServiceUpdate {
  title?: string;
  description?: string;
  icon?: string;
  features?: string[];
  order?: number;
}

export interface Message {
  id: string;
  name: string;
  email: string;
  /** @nullable */
  subject?: string | null;
  body: string;
  read: boolean;
  createdAt: string;
}

export interface MessageInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  email: string;
  subject?: string;
  /** @minLength 1 */
  body: string;
}

export interface MonthlyCount {
  month: string;
  count: number;
}

export interface CategoryCount {
  category: string;
  count: number;
}

export interface AdminStats {
  totalProjects: number;
  featuredProjects: number;
  totalBlogPosts: number;
  publishedBlogPosts: number;
  totalMessages: number;
  unreadMessages: number;
  totalServices: number;
}

export interface AdminAnalytics {
  messagesByMonth: MonthlyCount[];
  projectsByCategory: CategoryCount[];
  blogPostsByMonth: MonthlyCount[];
  totalContentPieces: number;
  responseRate: number;
}

export interface AdminLoginInput {
  /** @minLength 1 */
  password: string;
}

export interface AdminAuthResult {
  success: boolean;
  message: string;
}

export interface AdminSession {
  authenticated: boolean;
}

export type ListProjectsParams = {
featured?: boolean;
category?: string;
};

export type ListBlogPostsParams = {
published?: boolean;
};