# lib/api-zod/src/generated/types/blogPost.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * ajTech API specification
 * OpenAPI spec version: 0.1.0
 */

export interface BlogPost {
  id: string;
  title: string;
  slug: string;
  /** @nullable */
  excerpt?: string | null;
  content: string;
  /** @nullable */
  coverImageUrl?: string | null;
  tags: string[];
  published: boolean;
  /** @nullable */
  readTime?: number | null;
  createdAt: string;
  updatedAt: string;
}