# lib/api-zod/src/generated/types/message.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * ajTech API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Message {
  id: string;
  name: string;
  email: string;
  /** @nullable */
  subject?: string | null;
  body: string;
  read: boolean;
  createdAt: string;
}