# lib/api-zod/src/generated/types/messageInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * ajTech API specification
 * OpenAPI spec version: 0.1.0
 */

export interface MessageInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  email: string;
  subject?: string;
  /** @minLength 1 */
  body: string;
}